Use relative offsets instead of absolute pointers in shared memory
Summary:
Before this change, the hash table stored addresses as absolute pointers into
the heap. Because of this, it is necessary to map shared memory to a fixed
address, which is kind of yucky.

This diff replaces those absolute pointers with a relative pointer. Relative
pointers, addr_t, are offsets from the hashtbl base pointer shifted left by 1
position.

The left shift is necessary because the GC needs to distinguish addrs from
headers during its traversal. Pointers automatically had a 0 least-significant
bit because all heap entries are certainly 2-aligned (cache-aligned, actually).

The "null" address is represented as 0. This is possible because the base
pointer is the hashtbl start pointer. Thus, the address 0 points to the hash
field of the first entry in the hash table, which is a meaningless address and
will never appear. Since the minimum hash table size is 1 entry, this is always
okay.

0 is a convenient null address representation, because the empty hash table will
be initialized with 0, so all addresses will start out as null.

The "taken" address, used to snag a hash table entry during a write, is
represented as UINT64_MAX. Similarly, we will never need to point to this
address, which would represent an entry in a >18 exabyte heap. Note that I wrote
UINT64_MAX in "long hand" since many MSVC compilers don't support the C11
stdint.h definition.

Reviewed By: gabelevi

Differential Revision: D19183602

fbshipit-source-id: b4609b09a65bae7e20b7b12e76f201c51d15c5cf