upgrade to sedlex.2
Summary:
we have to upgrade to sedlex 2 to unblock using ppx_tools_versioned >= 5.2.1 which is necessary for newer versions of js_of_ocaml and other packages. it's also necessary for ocaml 4.08 compatibility.

2.0 switched from `int` to `uchar` so we have to work around that for unpaired surrogates which are legal in JS.

the new version also includes updated unicode, which is good because recent ecmascript specs (ES2016 IIRC) require using the latest version of unicode.

Reviewed By: gabelevi

Differential Revision: D18855252

fbshipit-source-id: 5479f92b78118fccab221608fbb54f0cb8b6d06e