better exception on write errors due to dead monitor
Summary:
when the monitor process exits without telling the server process[1], the server throws an EPIPE exception the next time the server tries to write a status update to the monitor's (closed) socket.

we can detect this and exit slightly more gracefully... it still essentially crashes the server, but now it does it with the correct exit code and message in the logs.

[1] I know of two cases when the monitor dies like this:

 - something `kill -9`s the monitor process, which doesn't give it a chance to tell the server
 - the monitor exits due to file watcher issues, since the monitor waits until the watcher is initialized before sending any messages to the server, including the shutdown message!

Reviewed By: gabelevi

Differential Revision: D19265027

fbshipit-source-id: 32bda9bb61a8a2aeef461d44cba9e23f130fb25c