object and instance access that seem likely to be typos now include a suggested property name in the error message
Summary:
This updates the `EStrictLookupFailed` and `EPropNotFound` error messages to include a possible suggested name change if the failed property access seems likely to be a typo. We determine if it's likely to be a typo by comparing the property we are looking for with all the properties in the object or instances inheritance hierarchy using the `typo_suggestion` function.

Note that this also adds a new field to `LookupT`: the set of property map ids we have already failed on in this lookup action. We need to do this because lookups walk the prototype chain, so we only know we have failed the lookup when we see a prototype, which itself has no fields on it.

Reviewed By: mroch

Differential Revision: D19047090

fbshipit-source-id: a56be5e50439ba4ce61a36d11cdff913a4b8d29f