Replace hand-rolled local_t alignment with C11 stdalign.h
Summary:
Pull Request resolved: https://github.com/facebook/flow/pull/8252

Shared memory stores worker-local data in the page(s) immediately following the
first "small objects" page. This is an array of local_t structs with one element
for each worker and one for the master process.

Before this diff, the structs were arranged manually, aligned to
CACHE_LINE_SIZE. Access was performed through a macro which handled the offsets.

Instead, we can use alignas, introduced in C11, to ensure that each local_t
struct is aligned to the cache line boundary. With this alignment in place, the
"standard" array access will just work, so the LOCAL macro is unnecessary.

Reviewed By: gabelevi

Differential Revision: D19268644

fbshipit-source-id: 3e057d6e57a1b4835c8cb268d19081f2704b8ccc