make uses of unknown_use generated by calls to rec_flow_t more explicit
Summary: We want to reduce the frequency with which we use this `use_op`, and it was created in many cases as the default value of calls to `rec_flow_t` that did not supply a `use_op` argument. By requiring calls to this function to supply this argument in every case, we make it easier to track down and replace improper uses of this `use_op`.

Reviewed By: panagosg7

Differential Revision: D19184897

fbshipit-source-id: ea132a93a1181b085fda0084acf32d78d6871c42