[normalizer] Fix bug in normalizer cache
Summary:
The normalizer uses a cache to store the result of normalized `OpenT`s and `EvalT`s.
Keys to this cache are just the identifiers of these types. What is not reflected in the
cache is whether the normalized version of the type includes any free recursive
variables.

To see why this is problematic, consider normalizing an `OpenT(i)` and `OpenT(j)`
such that
```
lower(i) = C<j>
lower(j) = D<i>
```
The steps for normalizing `OpenT(i)` (related to caching) are:

1. create fresh `I` for `i`
2. create fresh `J` for `j`
3. resolve `j` as `mu J . D<I>` and add to cache
4. resolve `i` as `mu I . C<mu J . D<I>>` and add to cache

The result for the top-level `i` is well-formed, but then if we wish to resolve `OpenT(j)`
as a top-level type, the cache will return
```
mu J . D<I>
```
which is now ill-formed, since `I` is out-of-scope.

This caused the following
```
const flattened = [[0, 1], [2, 3], [4, 5]].reduce(function(a, b) {
  return a.concat(b);
});
```
to return
```
Array<(any | number | V$1)> | Array<(any | number)>
```

This diff fixes this by skipping step (3) above. It does so by only caching results
if there are no `OpenT`s under normalization (quick check), or the result type does
not include free recursive variables (slower check).

Reviewed By: samwgoldman

Differential Revision: D19321547

fbshipit-source-id: 6405a6fd44181198bd58ae3ea9bd2061a7759eca