resolve type destructed unions after creation
Summary:
When we encounter a type destructor applied to a union type, we push the type destructor down into each element of the union to produce a union of eval types.

Normally, when we see a flow involving a lower bound union of eval types, we take the time to evaluate each of the components so that the union is able to hit our optimizations. However, the priority of the tvar rules is such that the result of the type destructor hits is stored in a tvar before its component eval types are resolved. Overall, the process looks something like this:
```
UnionT [X, Y, Z] ~> TypeDestructorTriggerT (f, tvar)
    UnionT [EvalT (f,X), EvalT (f, Y), EvalT (f, Z)] ~> tvar
```

This means that if the tvar ever appears as an upper bound, we are unable to take advantage of our speculation optimizations, as none of them are capable of peeking inside of unevaluated eval types.

Instead, we can flow the eval union produced by type destruction to a `ResolveUnionT`, and let the machinery that already exists for this purpose resolve the evals inside the union, and only then flow the resulting union to the tvar. This lets us hit our optimized checks.

This also changes the behavior of `$Diff` with respect to unions. Before, this was pretty broken, so the new behavior should be an improvement. The rules for the new behavior:

`$Diff<X | Y, Z> = X \ Z | Y \ Z`
 `$Diff<X, Y | Z> = X \ Y | X \ Z`

Reviewed By: panagosg7

Differential Revision: D18715235

fbshipit-source-id: 3a13059be991fb994f5d33193269cbe3dc771935