Drop shm_dirs config
Summary:
Pull Request resolved: https://github.com/facebook/flow/pull/8258

Flow spawns multiple processes which communicate via shared memory. To make this
work, the master process creates a shared file. Unfortunately, the various
platforms we support do not all support the same means for creating this shared
file.

Before this change, we tried one of four different methods:
* memfd_create, supported by Linux > 3.17
* shm_open, supported by BSD (OS X) and Linux
* a tmpfile in /dev/shm (or configured shm_dir)
* CreateFileMapping, a Windows-specific API

On Linux, we preferred memfd_create if available, or a tmpfile otherwise. We
avoid shm_open because the /dev/shm filesystem is often not large enough to
accommodate the requested size.

On BSD (OS X), we preferred shm_open if available, or a tmpfile otherwise.

On Windows, we always use CreateFileMapping.

This diff drops the tmpfile method. On Linux, if memfd_create is not available,
we will fall back to shm_open instead.

In the case where we would fall back to tmpfile, we will now fail to init. In
practice, I don't expect this will ever cause an issue.

The most likely failure scenario is where shm_open succeeds but ftruncate fails
to resize the file to the requested length. In this case, the solution is to
adjust the size of the shm file system.

I tried to search for any existing uses of the shm_dir and shm_min_avail
configuration, both internally and on GitHub, but could not find any existing
uses.

Reviewed By: mroch

Differential Revision: D19356583

fbshipit-source-id: 65363e76d457fb320cb32e57ee2d16c30bf122b4