Fix endless recheck loops
Summary:
Starting in Flow v0.115.0, we began to get reports of Flow getting stuck in a recheck cancellation loop. I believe that D18739041 is to blame: it caches promises (`Lwt.t`s), which can complete nondeterministically even if the underlying computation is deterministic. This happens when something external to the computation triggers a cancellation. At that point, `Lwt` bubbles up an `Lwt.Canceled` exception. Because we cache that promise, subsequent cache hits get canceled unintentionally, leading to endless retries.

This change makes the `Cache` data structure aware of `Lwt`. This way, it will cache results rather than the promises themselves.

Fixes #8259

Reviewed By: mroch, panagosg7

Differential Revision: D19435541

fbshipit-source-id: 3d6bf85b535065efb246147b956bf949c4fa93e8