kill the server when monitor can't start file watcher
Summary:
when we can't init watchman, we raise an exception which bubbled all the way up and fataled the monitor process. in practice, this is ok because we exit anyway when we can't init watchman.

handling it fixes the monitor process's exit code, improves logging, and properly notifies the server process to shut down instead of it just noticing that the monitor's pipe is unexpectedly closed. it'll also let us react better in the future (e.g. try watchman again).

Changelog: fix: Fixed `flow server` exit code when watchman can't be initialized

Reviewed By: avikchaudhuri

Differential Revision: D19431136

fbshipit-source-id: d6595eb5306ba5b99424f374d488ac074aeef374