Remove ty_loc from legacy autocomplete API
Summary:
the legacy autocomplete API would include a location with each completion result. This location wasn't being used in LSP. It was supposedly the "location of the definition of the type", which doesn't seem well-defined enough for anyone to have been using it for much (I hope).

My main motivation for finally removing this loc is that it would complicate moving autocomplete off of members.ml onto the ty_normalizer, since `Ty.t`s don't contain locs.

Reviewed By: nmote

Differential Revision: D18462303

fbshipit-source-id: 1810865342927e7e8c8ec6fffd2029084727762e