Autocomplete: use ty normalizer instead of members.ml
Summary:
Changes the implementation of autocomplete on member expressions to use the `Ty_normalizer` instead of `members.ml`. This is should be less bug-prone than the `members.ml` implementation because it gets to operate over the `Ty.t`s instead of `Type.t`s.

I've combed through the test changes here a few times and they all look reasonable to me, but it'd be great to get a second opinion on these changes in behavior.

Reviewed By: nmote

Differential Revision: D18914892

fbshipit-source-id: ca2900484f0c58ecbdb8aeda2720aa02fd5e3405