Suggest optional chaining when autocompleting a nullable object
Summary:
When an object was possibly null, autocomplete used to suggest its members as though it weren't possibly null, and so selecting one of the autocompletions would result in a type error. This diff changes that behavior so that when a user asks for completions on a possibly null/undefined object, the autocomplete suggestions include optional chaining so that selecting the completion doesn't result in a type error. This behavior was selected based on input we asked for from JS programmers.

The implementation works on a per-member basis, because it can be the case that some members need optional chaining and some don't: https://flow.org/try/#0PQKgBAAgZgNg9gdzCYAoVATApgYxgQwCcswA3IsADzAC4wB+Ab31rADsBXAWwCMtCANGB6sAzgBdCASzYBzAL5gAZGEYi6E6XPmpKAOnwBuXfQPH9PQ0A

Reviewed By: panagosg7

Differential Revision: D19420874

fbshipit-source-id: 08fce70bb7a66467a656fd07cb3c9ccf0e70a79c