Add static type to Ty.Fun
Summary: To use the ty normalizer instead of members.ml for member autocompletion, we need `Ty.t`s to store information about the members of a function type. That info is stored in the static portion of the function type, so this diff adds a field `fun_static` to function tys in which the static portion will be stored.

Reviewed By: panagosg7

Differential Revision: D18437995

fbshipit-source-id: 23fb597156b2997818eecc3c6cbc15d2d6d19c2d