Add expand_toplevel_members flag to ty_normalizer
Summary:
Adds an option `expand_toplevel_members` to `Ty_normalizer`. The spec of thie flag is roughly: make it so that it's possible to find out all of a type's members given the output from the normalizer. More concretely:
- it takes 2 arguments: the first says whether to include prototype members (turned off in JSX completion), and a hook for telling the caller whether the type is in an IdxWrapper.
- it expands type aliases, but differs from the `expand_type_aliases` flag in that we expand them only at the top level.
- it expands instance and class types into object types
- it expands primitive types into object types with their prototypes' members

The implementation accesses the state of this flag via `Ty_normalizer_env.expand_members env`, which returns an option. If at the current "level" the implementation should expand members, it returns `Some (info, env)` where `env` is the same environment as was passed in except with member expansion turned off. If members should continue to be expanded in recursive calls, `Ty_normalizer_env.continue_expanding_members env` is used to turn member expansion back on. I did it this way (defaulting to turning member expansion off after checking the flag, and needing to explicitly turn it back on) because
0. it makes more sense to have to explicitly specify when a recursive call should expand members than to explicitly specify when it shouldn't
0. I wanted potential bugs to err on the side of not expanding members in enough recursive calls, rather than expanding members too much in recursive calls. Not expanding enough in recursive calls would likely manifest as missing members (easier to notice), whereas expanding too much in recursive calls would likely manifest as showing the user strangely verbose/inaccurate types (harder to notice).

Reviewed By: panagosg7

Differential Revision: D18914882

fbshipit-source-id: 17e59e61cd2e7be2d95a7659298f3ac72db5f5b5