Autocomplete: rank proto props below own props
Summary: One thing that changed with the refactor of autocomplete off of `members.ml` is that now, object prototype members are included in the members of a class instance. This is consistent with a spec of "autocomplete should suggest everything that flow would accept here", but for better usability we'll rank the object prototype members below instance members.

Reviewed By: panagosg7

Differential Revision: D19319236

fbshipit-source-id: c49bc46d2b2555e75648a653992fbbf7f4a949f5