Add --lsp flag to autocomplete command for testing
Summary:
The existing LSP testing framework is admissive with respect to reordering strings in between commas in the responses, which makes it dicey at testing things like "does the JSON object whose 'label' property is 'baz' have a 'sortText' property of 0?"

This diff lets us test the LSP output of autocomplete more strongly by adding an `--lsp` flag to the `flow autocomplete` command. There's probably a better or more general way to get this sort of functionality for all LSP commands in the future, but for now this quick solution is useful for being able to test D19319236.

Reviewed By: nmote

Differential Revision: D19334179

fbshipit-source-id: 75847ae63a474acad4206a7e79a3bac7b20bf4ae