[watchman] incremental step toward handling failure to get mergebase
Summary:
we ask watchman to ask hg for the mergebase so we can watch all the files that changed locally, via `get_mergebase`. currently, if hg or watchman error, we `failwith` and die ungracefully.

ideally, we would handle the error and not die at all, perhaps by retrying (sometimes hg is in the middle of a rebase or something?), or perhaps just by sending the LSP a status message.

as a baby step, we should die gracefully, logging that it was a watchman issue and telling the server and monitor to both exit.

this diff is an even smaller baby step, which pushes the `failwith` further up the stack and uses `call_on_instance` to close our connection to watchman and update our internal state to know it's dead. we still just `failwith` right after that anyway, but followups will remove that.

Reviewed By: nmote

Differential Revision: D19520825

fbshipit-source-id: 01bd54b43f6d9bdaffd60e03c069f8577235fe91