pass file watcher timeout setting to server
Summary:
the new `--file-watcher-timeout` setting controls the monitor process's subscription to watchman. but during init in watchman lazy mode, the server process directly queries watchman for the changes since the mergebase, so it can focus them.

previously, we had a hardcoded timeout of 15 seconds longer than flow's init takes. that is, watchman gets to run in parallel during init, but we only let it block the critical path for 15 seconds.

this diff makes that setting also apply to the server's timeout.

Reviewed By: nmote

Differential Revision: D19624581

fbshipit-source-id: b573d3b82a274c0d8889fe5adc5c8528f0f2125e