get rid of default timeouts in watchman client
Summary:
we had a `Default_timeout` of 120s which was used on watchman requests. this is a footgun now that we're wrapping multiple requests into a single timeout, since that single timeout could be longer than the default timeout and then get capped if you forget to pass `~timeout:No_timeout`. so now `~timeout` is required.

I also removed `Default_timeout` entirely since it should be configurable. that only leaves `No_timeout` and `Explicit_timeout of float` so I changed it back to `float option`.

Reviewed By: gkz

Differential Revision: D19623337

fbshipit-source-id: 06f7e6f77a17a1d6000b5998f90584894d260393