move file watcher init timeout around entire init process
Summary:
before, there were many *hardcoded* timeouts when initializing watchman:

- 10s to get the socket name
- 120s to get watchman's capabilities
- 120s to call `watch-project`
- 30s to get the hg mergebase

the timeouts are there to make sure we don't get stuck, and there's some value in failing early if we know that some things should be fast. but the time it takes is related to the size of the project (and your machine), so they should be configurable, and it'd be a bad UX to add this many options.

so, here I remove all the individual timeouts and use a single timeout around the entire init process. it defaults to 120s. the next diff in the stack makes it configurable via flowconfig and cli. I think 120s is a pretty excessive default for most open source projects, so shouldn't cause issues.

Reviewed By: nmote

Differential Revision: D19588572

fbshipit-source-id: d4ae944ee2396c9a1c28df4198fff9258b309e85