[circle] don't run tests on tags
Summary:
before we push a tag, we make sure the tests are passing on the branch commit that we are about to tag. so running the tests again on the tag is redundant, slows down the deploy, and if the tests are flaky at all can cause the deploy automation to fail.

there are two ways we can make it not run tests:

1) a single Circle workflow, with filters making some jobs run on tags and branches, some run just on branches, and some run just on tags
2) two Circle workflows, one just for branches and the other just for tags

I went with (2) because I think it's clearer.

Reviewed By: jbrown215

Differential Revision: D19628444

fbshipit-source-id: 96be7276b4b604329efbfe404052e93a320b7c26