Remove outdated dependent files computation optimization
Summary:
This was added in D16432276, when computing dependents was an `O(dependency graph)` operation. However, D18580785 reduced the time complexity of this operation, and it's now much faster in practice.

In particular, it's implemented as a couple of folds over the input set. So, it's essentially a no-op in the case that this optimization hits. We probably save a few cycles by avoiding the additional function calls, but not enough to be worth the added complexity here (even as minimal as it is).

Reviewed By: gabelevi

Differential Revision: D19169345

fbshipit-source-id: 8cfb1410c0aef683865d9adb2439a88be018ac22