Assert that there are no dependencies in to_check
Summary:
D17821995 changes the computation of the `to_check` set to never put dependencies in it (when the flag, which was removed in D18051211, is on). Previously, we conflated `to_merge` and `to_check`, which meant that we had to make sure to drop the dependencies here.

It turns out that enforcing this invariant doesn't actually lead to much in terms of simplification opportunities, since we were decomposing the `to_check` set into `focused` and `dependents` anyway. Still, it's nice to enforce invariants and it does allow us to simplify one computation a little bit.

Reviewed By: gabelevi

Differential Revision: D18052046

fbshipit-source-id: e7316ee2ce3ca465c78e1cdc262df1b93e38c3f0