Store shmem layout in info page, break up memory mappings
Summary:
Pull Request resolved: https://github.com/facebook/flow/pull/8278

This diff changes the info page to include information about the layout of
shared memory, specifically the size of the locals, hash table, and heap
sections.

Since this information is available in the info page, we no longer need to pass
it out-of-band to workers in the heap handle record. Instead, the heap handle
can be just a file descriptor.

After this change, instead of creating a single memory mapping, processes will
have separate mappings for each of the info page, locals, and the hashtbl+heap.
Workers create the info page mapping first and use the layout information there
to create the other mappings.

Note that the hashtbl and heap are combined. This is necessary because they both
use relative offsets from the same base pointer.

Reviewed By: jbrown215

Differential Revision: D19320669

fbshipit-source-id: a27b341c88e580cff90f74f56e626e39152cdeae