[parser] add loc to arg list
Summary:
for signature help, we need to know if your cursor is within an argument list, so we need to know the location of the list.

we don't expose the loc in estree since it doesn't have a separate node for arg lists, so this change has no public API impact.

Reviewed By: gkz

Differential Revision: D18662334

fbshipit-source-id: 5fe7c2219f4d9d3a03dacda8cabf92bb42c80762