Add seed to Xx API
Summary:
Before this change, the C stubs would always just pass 0. This parameterizes the
API instead and updates callers to pass in 0 instead.

Even though the API takes a int64, the calls don't actually need to allocate a
block when compiling in native mode. The extern functions are already defined so
that these calls can be made without boxing. The only difference then is that
the appropriate register will be zeroed out before the call.

I'm going to use this later for hashing signatures.

Reviewed By: mroch

Differential Revision: D19783431

fbshipit-source-id: 7c6e2320deb73da138bc1aa9750186dbfe86578a