Always log time spent collating errors
Summary: This follows up on D19559108 by logging the time spent collating errors in its own category, instead of just counting it towards the total time of whatever calls it. This way we'll be able to see `CollateErrors` as a specific line item in the profiling data, which will allow us to have better insights into its performance in practice.

Reviewed By: mroch

Differential Revision: D19559113

fbshipit-source-id: 72b9fd43848794d1c1b85baad608e2fb8da95126