Log recheck event in rechecker.ml
Summary:
We need to include computations that happen outside of `types_js.ml` in this metric, so in preparation for this we need to move the logging up to the caller of `Types_js.recheck` in `rechecker.ml`.

The way I've chosen to do this is by partially applying the log function so that it only needs a `Profiling_js.finished`. This seems simpler than explicitly passing up all the various pieces of information that it requires.

Reviewed By: mroch

Differential Revision: D19559110

fbshipit-source-id: 80d1d974ed7aa06df895ecaa7903f6ce44c49107