Include error collation (when it happens) in recheck times
Summary:
Flow has to collate errors before sending them. Traditionally, this happened in response to the first `flow status` request after a recheck. These days, a lot more things can trigger error collation. In particular, when an LSP client is connected to Flow, we have to send errors down immediately after a recheck is complete. It makes sense to include this as part of the recheck time, since it is computation that has been triggered due to the recheck, it is on the critical path for getting a complete list of errors to the user, and it is not currently measured elsewhere.

To that end, I've moved the code that sends errors to persistent clients, including the code which collates errors, inside the function that gets timed by the call to `with_profiling_lwt`.

Reviewed By: mroch

Differential Revision: D19559108

fbshipit-source-id: 785f6af675804aca39e74a3a4ddda94bf49fe1d8