Move recheck summary computation into rechecker.ml
Summary:
This is just a simple refactor. The reason this is necessary is that in later diffs, a `Profiling_js.completed` will no longer be available in `Types_js.recheck`. Instead, its callers will pass down a `Profiling_js.running`.

Since the recheck summary needs the duration, it has to be computed elsewhere

Reviewed By: mroch

Differential Revision: D19559112

fbshipit-source-id: a1de1bec104ad2c5785ec8e87081b5bbcb4ef8c3