Remove unnecessary dependency calculation logic
Summary:
It struck me while reading this code that the `then` branch of this if statement should never be able to run. I'm not sure what its original purpose was, but it seems to have outlived its usefulness.

`sig_dependent_files` is the transitive closure of all dependent files, based on the `sig_dependency_graph`, from some starting point (what that starting point is isn't relevant for this argument). Each component is a strongly-connected component in the `sig_dependency_graph`. Therefore, each component is either completely inside of the set of `sig_dependent_files` or completely outside of it.

Reviewed By: samwgoldman

Differential Revision: D19917241

fbshipit-source-id: 74d75dab1230fa1a5b11b8c70e49d8dc56282892