Rename `preliminary_to_merge` and update comments
Summary:
I think `preliminary_dependencies` is a clearer name for this value. It's used as the starting point for the rest of the dependencies calculation, and that's the only way in which it influences the `to_merge` set that is ultimately computed with this information.

I also updated some comments.

Reviewed By: mroch

Differential Revision: D19917240

fbshipit-source-id: f43ce85f6e1046623124e8b0014ae8b1a09f2255