[normalizer] Only trust the toplevel RTypeAlias reason
Summary:
We use `RTypeAlias` to recover alias names and defining locations during
normalization. Until now, whenever we saw this reason description we
would look up the def_loc of the enclosing reason to determine the definition.

Problems:

1. Even though this definition is in general robust to reposition, there is no guarantee
that it will always be present: `Reason.def_poly_loc_of_reason` returns the
general location of the reason if the def_loc is missing.

2. For aliases whose body was an `ExactT` or `MaybeT` this reason description was
propagates to the internal portion. This might lead to better error messages, but
throws off the normalizer.

This change provides a more robust approach to address these two issues:

1. the defining location is in the reason description itself.
2. the defining location is only used for the toplevel constructor of `ExactT` or `MaybeT`.

Reviewed By: samwgoldman

Differential Revision: D20014665

fbshipit-source-id: 2d65baf9fef308d10a5c8de98f151742d434fe38