Update RTypeAlias reasons for ExactT and MaybeT in Flow_js
Summary:
When types of the form
```
ExactT t
MaybeT t
```
had an RTypeAlias on the top level, we used to replace this description on the
inner type `t` as well (recursively). This would happen in statement.ml, which was
a bit awkward and unintuitive. This change moves this to flow_js.ml.

Reviewed By: samwgoldman

Differential Revision: D19748069

fbshipit-source-id: e010c460c662224a6be4944913260a2af94b713b