Add reason to ShapeT
Summary:
This addresses a TODO in flow_js.ml around the shape type. There's no difference
in test behavior, but it simplifies things in the normalizer down the line, since we can
now inspect the reason of the Shape constructor itself, rather than the enclosing type
that we did before.

Reviewed By: samwgoldman

Differential Revision: D19781271

fbshipit-source-id: e1bc0f578dbffc89962465d28948f5d682b54a4d