[enums] Get def support for enum members
Summary: Add support for get-def on an enum member (e.g. the `A` in `E.A` if we have `enum E {A, B}`), going to the position of the enum member in the enum declaration. If we do get-def on one of the enum methods, to the lib def definition.

Reviewed By: vrama628

Differential Revision: D19980840

fbshipit-source-id: e7cb4b41bff2a5eef2d66dd18428ddde2b6b16c7