[enums] Refactor invalid member access error to take suggestion rather than all members
Summary: Rather than have the error take in the set of all members, have it take the suggestion instead. This matches how dsainati1 did the errors for prop not found, and helps me refactor the members from a set to a map (which I need to do in the future). (Might also help vrama628 in adding fix it suggestion for both this and prop not found).

Reviewed By: vrama628

Differential Revision: D19980843

fbshipit-source-id: f96c378495ce3e5e9e840aba0746c25cb2d13342