[enums] Autocomplete support for enum members and methods
Summary:
Adds support for autocompleting enum members, and methods on the prototype. The prototype methods are sorted after the enum members. We don't have Object.prototype on the prototype chain, so should not see things such as `.toString`

We use `Env.expand_primitive_members` to sort the prototype methods below the enum methods. I would like to make a follow up diff to this one renaming this to something more appropriate. Suggestions?

Reviewed By: vrama628

Differential Revision: D20014042

fbshipit-source-id: 9f37195659476a2af8d345700bb7233e895e7f24