[normalizer] Force declaration types to be toplevel only (part 4)
Summary:
## Motivation

See {D20070324}

## This diff

Main changes:

* Forces the class declaration constructor (Ty.ClassDecl) to only appear at the toplevel.
* When a Type.ClassT appears nested in a type it will correspond to the `Class<T>`
utility.
* An inline Type.ThisClassT corresponds to `typeof C`.

Note that a toplevel `ThisClassT` normalizes to the type Ty.ClassDeclaration rather than
a Ty.TypeOf type. This will be easy to adjust later on, if for example we really want all
expressions to get proper normalized types, instead of type declarations.

Minor changes:

TypeAlias is used both for
```
type A = ...
```
and for
```
import { type A as B } from '...';
              ^
```
I added a field to distinguish these cases. It's useful to know when this is a case of import,
to figure out which exactly remote symbol is imported (see Import.extract_ident).

Reviewed By: samwgoldman

Differential Revision: D20127817

fbshipit-source-id: 18e6f07c117cbf25b0410a50765aea2b68c15347