[normalizer] Force declaration types to be toplevel only (part 2)
Summary:
## Motivation

See {D20070324}

## This diff

The main change here is the removal of the recursive call to `type_t` in `poly_ty`,
which allows us to pull `type_t` out of the main recursion.

In order to make this happen I had to make the following changes:

1/ Change an awkward pattern used in side A of `type_app` and `type_t`:

Before, in order to normalize `TypeAppT (t, ts)`, we would first normalize `t` into
`ty` and then pattern match on `ty` to figure out the end result by applying the arguments
`ts`. This pattern  won't be an option anymore, because soon *some* of the intermediate
`ty` will not be Ty.t  elements, but declarations. The latter would require us to call `type_t`
(for example) in the same recursion, which is exactly the pattern we're trying to get rid of.

This diff bypasses this by inspecting the input Type.t types directly.

2/ Enhanced the "under_type_alias" environment.

As the normalizer recurses into various type structures, it needs to keep track of the
aliases it has seen to avoid infinitely expanding type aliases. Before this diff, we relied
on a set of "seen" OpenTs to avoid the infinite recursion. Now that we make aggressive
use of the Lookahead mechanism, we might bypass some of these OpenTs. To maintain
termination we keep a set of the symbols we've seen and check for them in
`type_with_alias_reason`.

3/ Added a mechanism to keep track of the entry callback to the `type__` recursion.

Once we've decided that an alias is to be expanded we move from `type_with_alias_reason`
that we saw before and into `type_after_reason`. In the case of EvalTs, in particular, we
might additionally need to run code in Flow_js and normalize the result. Instead of
calling recursively into this process through `type__`, we'd like to return back to
`type_after_reason` (ie. bypass the type alias shortcut). This is where the `~cont`
argument comes in handy.

It's possible that we could have reused the existing machinery, by updating the
environment with some relevant information, but this option feels slightly simpler.

Reviewed By: samwgoldman

Differential Revision: D20082709

fbshipit-source-id: e53fe28c81404ce0b596d986a2811fb1ede6e75b